Added appendable feedback
This mechanism allows you to add new feedback depending on whether the "Last Updated" timestamp matches the timestamp of the feedback given. This way you can edit feedback given as long as you are seeing the same ticket.

This is currently based on "createdAt" timestamp.

IF WE WANT TO ALLOW FEEDBACK FOR DIFFERENT ITERATIONS, WE NEED TO SWITCH THE TIMESTAMP TO LAST TIME THE SUMMARY WAS UPDATED FOR THE TICKET.